+++ src/types/form/signup.ts
@@ -1,0 +1,8 @@
+export interface SignupFormValues {
+  email: string;
+  password: string;
+  passwordConfirm: string;
+  nickname: string;
+  privacy: boolean;
+  marketing: boolean;
+}
